refactor(windows): move the common functions into connlib_shared::windows
