+++ rust/connlib/shared/Cargo.toml
@@ -38,6 +38,7 @@ hickory-resolver = { workspace = true }
 domain = { workspace = true }
 libc = "0.2"
 dns-lookup = { workspace = true }
+known-folders = "1.1.0"
 
 # Needed for Android logging until tracing is working
 log = "0.4"
