--- rust/windows-client/firezone-windows-common/Cargo.toml
@@ -1,7 +1,0 @@
-[package]
-name = "firezone-windows-common"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-known-folders = "1.1.0"
