+++ rust/windows-client/src-tauri/src/client/debug_commands.rs
@@ -45,7 +45,7 @@ fn hostname() -> Result<()> {
 /// Try to load wintun.dll and throw an error if it's not in the right place
 fn wintun() -> Result<()> {
     tracing_subscriber::fmt::init();
-    let path = firezone_windows_common::wintun_dll_path()
+    let path = connlib_shared::windows::wintun_dll_path()
         .ok_or_else(|| anyhow!("can't compute wintun_dll_path"))?;
     unsafe { wintun::load_from_path(&path) }?;
     tracing::info!(?path, "Loaded wintun.dll");
