+++ rust/windows-client/src-tauri/src/client/deep_link.rs
@@ -2,7 +2,7 @@
 //! Based on reading some of the Windows code from <https://github.com/FabianLars/tauri-plugin-deep-link>, which is licensed "MIT OR Apache-2.0"
 
 use crate::client::auth::Response as AuthResponse;
-use firezone_windows_common::BUNDLE_ID;
+use connlib_shared::windows::BUNDLE_ID;
 use secrecy::SecretString;
 use std::{ffi::c_void, io, path::Path};
 use tokio::{io::AsyncReadExt, io::AsyncWriteExt, net::windows::named_pipe};
