+++ rust/windows-client/src-tauri/src/client/gui.rs
@@ -10,8 +10,7 @@ use crate::client::{
 use anyhow::{anyhow, bail, Context, Result};
 use arc_swap::ArcSwap;
 use connlib_client_shared::{file_logger, ResourceDescription};
-use connlib_shared::messages::ResourceId;
-use firezone_windows_common::BUNDLE_ID;
+use connlib_shared::{messages::ResourceId, windows::BUNDLE_ID};
 use secrecy::{ExposeSecret, SecretString};
 use std::{net::IpAddr, path::PathBuf, str::FromStr, sync::Arc, time::Duration};
 use system_tray_menu::Event as TrayMenuEvent;
